Renamed Models folder to models in objects and unit_tests folders, and made private variables in api_calls model
